copied over reverse_stitching stuff
