+++ cameraConfigs.cfg
@@ -1,0 +1,52 @@
+1.0
+1357.382668795922
+1030.2095730373799
+1814.3069109793582
+-0.21823664009571075
+0.9755886793136597
+-0.024483606219291687
+0.02435777522623539
+-0.019635174423456192
+-0.9995105266571045
+-0.975591778755188
+-0.21872617304325104
+-0.019478056579828262
+0.0
+0.0
+0.0
+-4373
+-4957
+1.0
+1448.9221729727672
+928.5284127544213
+2234.0940125814873
+-0.23855572938919067
+0.9705901145935059
+-0.03234099596738815
+-0.02438271790742874
+-0.039278216660022736
+-0.9989308714866638
+-0.9708226323127747
+-0.23751211166381836
+0.03303566575050354
+0.0
+0.0
+0.0
+-4356
+-3865
+1.0
+758.5593830446072
+458.0
+327.5
+0.9867041110992432
+0.14352406561374664
+0.07626182585954666
+-0.0012990031391382217
+0.47617489099502563
+-0.8793497681617737
+-0.16252180933952332
+0.8675587773323059
+0.4700300693511963
+0.0
+0.0
+0.0
